Support skipping upload if file already exists (#201)
## Background

Currently, files are always uploaded to the remote side, even if they
didn't change between runs. This causes new versions to be created,
incurring potentially unneeded costs and resource waste.

In some cases overwrite might be desired, e.g. for backups, as the files
might have changed on disk.
In other scenarios, the CI might be building idempotent assets, in which
case "overwrite" is not desired, and the upload should be skipped.

The default will still be `overwrite=true`, i.e. it won't break any
workflow in use.

## Solution

Add param `overwrite=true/false` which skips the upload in case the file
already exists. This saves bandwith and storage space on the remote
side.

## Implementation

Before doing anything, a call to `HeadObjectInput` is done which checks
if the file already exists. If so, the upload is skipped.

The implementation could potentially be improved by checking the sha of
the object on both sides and only skip the upload if both are identical.
The current solution is sufficient for my use case, though.

## Testing

I am using a custom image with this branch since weeks already - didn't
encounter any issues so far.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>